Refactor to parse into collection
This makes a Collection (a set of EnvVar objects) the default
result of parsing, with all other formats (hash, string, array of
strings) defined on the Collection class.

This keeps the same behaviour that we have now, but allows
for the Collection to be used when merging env vars.

The next step would be to define SecureEnvVar, also as a set member
class (#eql?, #hash) which would allow collections of
secure and non-secure env vars to be merged.

Also removes the merge mode methods, which can't be defined here as
this parser will never operate on the full config hash.